Install eslint-plugin-react-hooks
