+++ client/package.json
@@ -2,6 +2,16 @@
   "name": "client",
   "version": "0.1.0",
   "private": true,
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "lint": "eslint src/**/*.{ts,tsx}",
+    "lint:fix": "eslint -- --fix",
+    "lint:pr": "eslint src/**/*.{ts,tsx} --fix-dry-run --max-warnings 0",
+    "prettier": "prettier --write \"**/*.{ts,tsx,json,scss,md}\"",
+    "generate": "graphql-codegen"
+  },
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
@@ -22,49 +32,22 @@
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.1",
     "recharts": "^2.1.10",
     "subscriptions-transport-ws": "^0.11.0",
     "typescript": "^4.6.3",
     "urql": "^2.2.0",
     "web-vitals": "^2.1.4"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "lint": "eslint src/**/*.{ts,tsx}",
-    "lint:fix": "eslint -- --fix",
-    "lint:pr": "eslint src/**/*.{ts,tsx} --fix-dry-run --max-warnings 0",
-    "prettier": "prettier --write \"**/*.{ts,tsx,json,scss,md}\"",
-    "generate": "graphql-codegen"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^5.21.0",
     "@typescript-eslint/parser": "^5.21.0",
+    "@vitejs/plugin-react": "^1.3.0",
     "autoprefixer": "^10.4.7",
     "eslint": "^8.14.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
     "postcss": "^8.4.13",
     "prettier": "2.6.2",
-    "tailwindcss": "^3.0.24"
+    "tailwindcss": "^3.0.24",
+    "vite": "^2.9.9"
   }
 }
